% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{earthlinks_ui}
\alias{earthlinks_ui}
\alias{earthlinks_server}
\title{EarthLinks main modules}
\usage{
earthlinks_ui()

earthlinks_server(input, output, session)
}
\arguments{
\item{input, output, session}{Internal shiny arguments.}
}
\value{
UI and server components that can be used as arguments for
\code{\link[shiny]{shinyApp}}.
}
\description{
Create the UI and server function of the EarthLinks shiny app. Useful
if you need to embed EarthLinks into your own shiny app.
}
\examples{
shiny::shinyApp(ui = earthlinks_ui, server = earthlinks_server)
}
