% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run EarthLinks}
\usage{
run_app(options = list())
}
\arguments{
\item{options}{See \code{\link[shiny]{shinyApp}}.}
}
\value{
A shiny app.
}
\description{
Run EarthLinks
}
\examples{
run_app()
}
